mainly tidying up code

+++ src/Program.cs
@@ -2,7 +2,7 @@ class Program
 {
     private static void Main()
     {
-        int maxPlayers = 10;
+        byte maxPlayers = 10;
         int port = 1943;
         Server server = new Server();
         server.StartServer(maxPlayers, port);
